+++ CHANGES.rst
@@ -4,6 +4,12 @@ ThreeDiToolBox changelog
 
 1.4 (unreleased)
 ----------------
+
+- Gracefully handle HTTPError thrown by ResultsWorker thread.
+
+- Added QSortFilterProxyModel to the result_selection to enable sorting and
+  filtering of downloaded results.
+
 - WaterbalanceTool account for flow directions 1d2d, 1d, 2d and groundwater
 
 - Enable leakage and simple infiltration through aggregation NetCDF